Refactor: Remove getOS function
The getOS function, which was used to detect the operating system, has been removed from the codebase. This function was previously using the navigator.userAgent property to determine the OS. However, it is no longer needed and has been deleted.

This commit refactors the code by removing the getOS function entirely.

Ref: Issue #123